refactor: decouple cli functions into function mod
