+++ scripts/build_iOS_lib.sh
@@ -18,8 +18,4 @@ mkdir output/ios/proto/sign
 mv ./target-mobile/MaskWalletCoreMobile.h ./output/ios/MaskWalletCoreMobile.h
 cp ./target/universal/release/libmask_wallet_core_mobile.a ./output/ios/
 
-pushd chain-common/proto
-protoc --swift_opt=Visibility=Public --swift_out=../../output/ios/proto ./*.proto
-
-pushd sign
-protoc --swift_opt=Visibility=Public --swift_out=../../../output/ios/proto/sign ./*.proto
+sh ./scripts/proto.sh
